+++ src/frontend/app/sdoc/components/sdoc-searchpage/sdoc-searchpage.component.ts
@@ -341,6 +341,10 @@ export class SDocSearchpageComponent implements OnInit, OnDestroy {
         return false;
     }
 
+    onScrollToTop() {
+        this.pageUtils.scrollToTop();
+    }
+
     private redirectToSearch() {
         // reset initialized
         this.initialized = false;